Add files via upload
Spotify is a Swedish audio streaming and media services provider founded in 2006 by Daniel Ek. It is one of the world's largest music streaming service providers, with over 356 million monthly active users, including 158 million paying subscribers, as of March 2021.

So in this project a recommendation system is developed using unsupervised learning techniques.